backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector Store & Embeddings
chromadb==0.4.15
openai==1.3.0

# LangChain
langchain==0.1.0
langchain-openai==0.0.2

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.0
flake8==6.1.0
isort==5.13.2
